backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
httpx>=0.27.0
python-multipart>=0.0.9
cryptography>=43.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=3.0.3
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
tenacity>=8.3.0
python-slugify>=8.0.4
beautifulsoup4>=4.12.3
